Removing loading screen
